pyproject.toml
[project]
name = "spark-lakehouse"
version = "0.1.0"
description = ""
authors = [
    {name = "arempter",email = "arempter@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyspark (>=4.0.0,<5.0.0)",
    "delta-spark (==4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
